+++ example/lib/extensions/numeric.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+extension Onint on int {
+  SizedBox get vertical => SizedBox(height: toDouble());
+}