+++ lib/src/instaservices.dart
@@ -51,15 +51,9 @@ class Instaservices {
   Future<String> getUsername({
     required String accesstoken,
     required String userid,
-    List<String>? fields,
   }) async {
     try {
-      String scopes = '';
-      if (fields != null) {
-        scopes = fields.join(',');
-      } else {
-        scopes = usernameFields.join(',');
-      }
+      String scopes = usernameFields.join(',');
       final response = await http.get(
         Uri.parse(
           'https://graph.instagram.com/$userid?fields=$scopes&access_token=$accesstoken',
@@ -71,6 +65,50 @@ class Instaservices {
     } catch (e) {
       debugPrint('Error: $e');
       throw Exception(e);
+    }
+  }
+
+  Future<Map<String, dynamic>?> getContent({
+    required String accesstoken,
+    required String userid,
+    List<String>? fields,
+  }) async {
+    try {
+      String scopes = '';
+      if (fields != null) {
+        scopes = fields.join(',');
+      } else {
+        scopes = 'media_count,account_type';
+      }
+      final response = await http.get(
+        Uri.parse(
+          'https://graph.instagram.com/$userid?fields=$scopes&access_token=$accesstoken',
+        ),
+      );
+      final body = jsonDecode(response.body);
+      debugPrint('body: $body');
+      return body;
+    } catch (e) {
+      debugPrint('Error: $e');
+      throw Exception(e);
+    }
+  }
+
+  Future<List<dynamic>> fetchUserMedia({
+    required String userId,
+    required String accessToken,
+  }) async {
+    final response = await http.get(
+      Uri.parse(
+          'https://graph.instagram.com/$userId/media?fields=id,media_type,media_url,timestamp&access_token=$accessToken'),
+    );
+
+    if (response.statusCode == 200) {
+      Map<String, dynamic> data = json.decode(response.body);
+      debugPrint('Data: $data');
+      return data['data']; // 'data' is the list of media items
+    } else {
+      throw Exception('Failed to load media data');
     }
   }
 }